requirements.txt

fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pyjwt
email-validator
